refactor: rename painter to canvas
